# faster pop_tz_offset_from_string, 2x faster dateparser.parse if date has no timezone info