--- a/dateparser/timezone_parser.py
+++ b/dateparser/timezone_parser.py
@@ -33,7 +33,7 @@ class StaticTzInfo(tzinfo):
 
 
 def pop_tz_offset_from_string(date_string, as_offset=True):
-    if _search_regex.search(date_string):
+    if _search_regex_ignorecase.search(date_string):
         for name, info in _tz_offsets:
             timezone_re = info['regex']
             timezone_match = timezone_re.search(date_string)
@@ -87,4 +87,6 @@ def get_local_tz_offset():
 _search_regex_parts = []
 _tz_offsets = list(build_tz_offsets(_search_regex_parts))
 _search_regex = re.compile('|'.join(_search_regex_parts))
+_search_regex_ignorecase = re.compile(
+    '|'.join(_search_regex_parts), re.IGNORECASE)
 local_tz_offset = get_local_tz_offset()