When most strings don't have tz offset, this is massively faster, as we avoid a loop over all timezones (around 800 of them).
But it's possible to improve this.

TODO:

- [x] benchmark results
- [ ] improve case when we do find a match