@lopuhin 2019-09-23
Tests run quite a bit faster after this change (e.g. ``tests/test_date_parser.py`` went from 1.7s to 1.5s, ``tests/test_search.py`` went from 2.2s to 1.9s), also I wrote a little benchmark specifically for this PR:

```
$ cat t.py 
import time
import random
import dateparser

def main():
    inputs_no_tz = []
    inputs_tz = []
    random.seed(42)
    for _ in range(1000):
        day = random.randint(10, 31)
        year = random.randint(1970, 2030)
        hour = random.randint(12, 23)
        date = f'{year}.12.{day} {hour}:45:00'
        inputs_no_tz.append(date)
        inputs_tz.append(date + ' EST')

    t0 = time.perf_counter()
    for x in inputs_tz:
        assert dateparser.parse(x)
    t_tz = (time.perf_counter() - t0) / len(inputs_tz)

    t0 = time.perf_counter()
    for x in inputs_no_tz:
        assert dateparser.parse(x)
    t_no_tz = (time.perf_counter() - t0) / len(inputs_no_tz)

    print(f'Time with tz: {t_tz:.6f} s, t_no_tz: {t_no_tz:.6f} s')


if __name__ == '__main__':
    main()
```

And here are results (on branch first):

```
$ python t.py
Time with tz: 0.001375 s, t_no_tz: 0.000579 s
$ python t.py
Time with tz: 0.001363 s, t_no_tz: 0.000578 s
$ git checkout master
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
$ python t.py
Time with tz: 0.001353 s, t_no_tz: 0.001223 s
$ python t.py
Time with tz: 0.001344 s, t_no_tz: 0.001216 s
```
so parsing without TZ is more than 2x faster, while with TZ it's a tiny bit smaller. It should also be possible to bring the time of TZ parsing down.