@lopuhin 2019-09-23
@asadurski do you think it's fine to merge this PR as it is now (big speedup for dates without TZ, small slowdown for dates with TZ), or it's better to improve the case with TZ as well? It should be possible to eliminate the loop from ``pop_tz_offset_from_string`` entirely, but it would be a bit more changes than now.