@lopuhin 2019-09-23
I also checked import time, and it's affected a little as well:
```
# PR
$ for i in $(seq 10); do python -c "import time; t0 = time.perf_counter(); import dateparser; print(f'{(time.perf_counter() - t0):.3f} s')"; done
0.126 s
0.128 s
0.129 s
0.127 s
0.129 s
0.126 s
0.130 s
0.127 s
0.126 s
0.126 s
# master
$ for i in $(seq 10); do python -c "import time; t0 = time.perf_counter(); import dateparser; print(f'{(time.perf_counter() - t0):.3f} s')"; done
0.114 s
0.115 s
0.112 s
0.112 s
0.112 s
0.112 s
0.111 s
0.112 s
0.112 s
0.111 s
```