@lopuhin 2019-09-23
> I also checked import time, and it's affected a little as well:

It should be possible to reverse this though if we lazy-init stuff used by timezone parser.