@lopuhin 2019-09-26
Thank you @asadurski 👍 

> And I'll look some more into performance benchmark options with Travis.

I think even a benchmark suite which can be run locally to compare with master would be very useful, and there are plenty of nice cases in the tests which could be used as benchmark input.